deep_face/app/requirements.txt
tf-keras
deepface
opencv-python
uvicorn[standard]
pillow
numpy
python-multipart
fastapi==0.98.0
